+++ policyengine_canada/variables/gov/provinces/on/benefits/on_benefits.py
@@ -1,0 +1,12 @@
+from policyengine_canada.model_api import *
+
+
+class on_benefits(Variable):
+    value_type = float
+    entity = Household
+    label = "ON benefits"
+    documentation = "Ontario benefits"
+    unit = CAD
+    definition_period = YEAR
+
+    adds = "gov.province.on.tax.income.credits.refundable"