+++ policyengine_canada/variables/household/expenses/childcare/childcare_received_days.py
@@ -1,0 +1,9 @@
+from policyengine_canada.model_api import *
+
+
+class childcare_received_days(Variable):
+    value_type = int
+    entity = Household
+    label = "Received days of child care in a year"
+    unit = "day"
+    definition_period = YEAR